Fixed 2 bugs
Bug that prevented recipe list recycler view from
displaying the names of recipes was fixed. The problem
was simply that the Name variable in the Recipe class
named to be all lower case. Another bug that prevented
the recycler view items from being clicked was fixed.
The issue was that the onClickListener was not being
set on to the view holder in the constructor.