Finished basic app functionality
Bugs worked out so that all navigation features are
currently working in the app. In addition, data is
correctly being gathered and connected to the UI. The
only data to UI element left is the videos. UI still
needs polish, and widget still needs to be made. Also
network problems still need to be handled.