Fixed video bug with next and previous buttons
When either the next or previous button was pressed
in StepDetailFragment, all videos would stop working.
The fix was to simply delete a player release call
in the non video case when the data source changed.

Also, a player pause call was added to non video steps
in order to stop the video audio still playing while
viewing steps without videos.