Completed various tasks related to Widget GridView
Added two helper functions to the ListView's adapter
to process the recipe index and ingredient items.
Removed the title TextView from the widget layout. This
combined with the processIndex helper function in the
adapter means that the RecipeWidgetProvider class will
not need to receive any information back from the
adapter. Also added the ListWidgetService to the app
manifest. Also added a recipeIndex member variable to
the adapter.