Data now displays in widget
I initially tried to reference the mRecipes list
in ListWidgetService before it was assigned a value.
In addition, I was using an unsupported view, the
CardView. Fixing both of these issues resulted in
ingredient data properly displaying in the widget.
The next step is to add functionality to the buttons.